ch4/ucx: move multivci code to ucx_vci.c
